IntoDigits: replace `sort` with `sort_unstable` in permutations methods.
Since the initial order is irrelevant, `sort_unstable` is preferable since it is slightly faster and sorts in place.

The `is_binary_permutation` now relies on the number of 1's in each number, as the number of 0's is irrelevant (left-padding can add as many as needed). Also, since it no longer relies on a `Vec` for the comparison, it can now be used without the standard library.